refactor(ui): revert to using single tree for form data

+++ invokeai/frontend/web/src/features/nodes/components/sidePanel/builder/FormElementEditModeHeader.tsx
@@ -4,22 +4,13 @@ import { useAppDispatch } from 'app/store/storeHooks';
 import { ContainerElementSettings } from 'features/nodes/components/sidePanel/builder/ContainerElementSettings';
 import { useDepthContext } from 'features/nodes/components/sidePanel/builder/contexts';
 import { NodeFieldElementSettings } from 'features/nodes/components/sidePanel/builder/NodeFieldElementSettings';
+import { useIsRootElement } from 'features/nodes/components/sidePanel/builder/shared';
 import { formElementRemoved } from 'features/nodes/store/workflowSlice';
 import { type FormElement, isContainerElement, isNodeFieldElement } from 'features/nodes/types/workflow';
 import { startCase } from 'lodash-es';
-import { memo, useCallback } from 'react';
+import { memo, useCallback, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { PiXBold } from 'react-icons/pi';
-
-const getHeaderLabel = (el: FormElement) => {
-  if (isContainerElement(el)) {
-    if (el.data.layout === 'column') {
-      return 'Container (column layout)';
-    }
-    return 'Container (row layout)';
-  }
-  return startCase(el.type);
-};
 
 const sx: SystemStyleObject = {
   w: 'full',
@@ -44,28 +35,44 @@ export const FormElementEditModeHeader = memo(
     const { t } = useTranslation();
     const depth = useDepthContext();
     const dispatch = useAppDispatch();
+    const isRootElement = useIsRootElement(element.id);
     const removeElement = useCallback(() => {
+      if (isRootElement) {
+        return;
+      }
       dispatch(formElementRemoved({ id: element.id }));
-    }, [dispatch, element.id]);
+    }, [dispatch, element.id, isRootElement]);
+    const label = useMemo(() => {
+      if (isContainerElement(element)) {
+        const baseLabel = isRootElement ? 'Root Container' : 'Container';
+        if (element.data.layout === 'column') {
+          return `${baseLabel} (column layout)`;
+        }
+        return `${baseLabel} (row layout)`;
+      }
+      return startCase(element.type);
+    }, [element, isRootElement]);
 
     return (
       <Flex ref={ref} sx={sx} data-depth={depth}>
         <Text fontWeight="semibold" noOfLines={1} wordBreak="break-all">
-          {getHeaderLabel(element)}
+          {label}
         </Text>
         <Spacer />
         {isContainerElement(element) && <ContainerElementSettings element={element} />}
         {isNodeFieldElement(element) && <NodeFieldElementSettings element={element} />}
-        <IconButton
-          tooltip={t('common.delete')}
-          aria-label={t('common.delete')}
-          onClick={removeElement}
-          icon={<PiXBold />}
-          variant="link"
-          size="sm"
-          alignSelf="stretch"
-          colorScheme="error"
-        />
+        {!isRootElement && (
+          <IconButton
+            tooltip={t('common.delete')}
+            aria-label={t('common.delete')}
+            onClick={removeElement}
+            icon={<PiXBold />}
+            variant="link"
+            size="sm"
+            alignSelf="stretch"
+            colorScheme="error"
+          />
+        )}
       </Flex>
     );
   })
